gym-pyro-master/requirements.txt
gym>=0.17.1
numpy>=1.18.1
ply>=3.11
pyro-ppl>=1.2.1
torch>=1.4.0
